fix scaling of texture coordinates for previewsurface
The issue was introduced with commit ba61426ade3383119aa3cffa2c0ed21d40ebfa58. Correct scaling requires that ospray texture map scale is set to the reciprocal of scale factor from Transform2dNode. Also it seemed not to be correct to take scale factor from UsdUVTexture as scale factor for texture map. Instead the scale factor from UsdUVTexture is meant to be applied to the values which are read from the texture.